refactor: 하려다가 실패한 코드, 기록용 커밋
refactor: #67
